requirements.txt
fastapi
uvicorn[standard]>=0.30
python-dotenv>=1.0.1
openai>=1.40.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
chromadb>=0.5.0
